Restructure packages and update imports for handler and simplex
