+++ backend/internal/handler/handler_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"bytes"
