+++ backend/internal/handler/validations.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"fmt"
