+++ backend/internal/simplex/simplex_test.go
@@ -1,4 +1,4 @@
-package main
+package simplex
 
 import (
 	"testing"
